chore: Move manifest tests to common
